change dir of ExtrinsicObject
