+++ rust/src/ffi/util.rs
@@ -2,12 +2,12 @@ use std::any;
 use std::any::TypeId;
 use std::collections::{HashMap, HashSet};
 use std::convert::TryFrom;
-use std::ffi::CString;
+use std::ffi::{CString, c_void};
 use std::ffi::{CStr, IntoStringError, NulError};
 use std::os::raw::c_char;
 use std::str::Utf8Error;
 
-use crate::domains::ffi::{ExtrinsicObject, ExtrinsicDomain};
+use crate::domains::ffi::ExtrinsicDomain;
 use crate::domains::{AtomDomain, OptionDomain, VectorDomain};
 use crate::error::*;
 use crate::ffi::any::{AnyObject, AnyQueryable};
@@ -33,6 +33,27 @@ use std::marker::PhantomData;
 pub struct Sequential<T>(PhantomData<T>);
 #[cfg(not(feature = "untrusted"))]
 pub struct Pairwise<T>(PhantomData<T>);
+
+pub type RefCountFn = extern "C" fn(*const c_void, bool) -> bool;
+
+#[repr(C)]
+pub struct ExtrinsicObject {
+    pub(crate) ptr: *const c_void,
+    pub(crate) count: RefCountFn,
+}
+
+impl Clone for ExtrinsicObject {
+    fn clone(&self) -> Self {
+        (self.count)(self.ptr, true);
+        Self { ptr: self.ptr.clone(), count: self.count.clone() }
+    }
+}
+
+impl Drop for ExtrinsicObject {
+    fn drop(&mut self) {
+        (self.count)(self.ptr, false);
+    }
+}
 
 #[derive(Debug, PartialEq, Clone)]
 pub enum TypeContents {
